Address the floor from the extract package, not the XML module
Both gates read "from ...etl.extract.xml import MIN_FULLTEXT_CHARS" on the
line above a check applied to HTML output. The constant stays defined
beside MAX_STUB_NOTICE_CHARS, which is where its reasoning belongs, but it
is re-exported from etl.extract so the import reads like what it is: an
extraction-layer floor, not an XML one.

_extract_scope also now says outright who it is for - callers that own the
tree - so reaching for it instead of the public method reads as opting out
of the copy rather than as picking the shorter name.
