+++ src/linkml_reference_validator/etl/extract/__init__.py
@@ -4,10 +4,16 @@ from linkml_reference_validator.etl.extract.base import Extractor, ExtractorRegi
 
 # Import extractors to register them
 from linkml_reference_validator.etl.extract.html import HTMLExtractor
-from linkml_reference_validator.etl.extract.xml import XMLExtractor
+from linkml_reference_validator.etl.extract.xml import (
+    MAX_STUB_NOTICE_CHARS,
+    MIN_FULLTEXT_CHARS,
+    XMLExtractor,
+)
 from linkml_reference_validator.etl.extract.pdf import PDFExtractor
 
 __all__ = [
+    "MAX_STUB_NOTICE_CHARS",
+    "MIN_FULLTEXT_CHARS",
     "Extractor",
     "ExtractorRegistry",
     "HTMLExtractor",
